Revert "Merge pull request #88 from RusticLuftig/index-compression"
This reverts commit 41745e56c6b3442d04f06dde874706f7e3c6e641, reversing
changes made to 161565de8066de9c8f78d0406e9bcfa03a9cf28a.
