+++ common/constants.py
@@ -8,9 +8,3 @@ DATA_ENTITY_BUCKET_SIZE_LIMIT_BYTES = utils.mb_to_bytes(128)
 DATA_ENTITY_BUCKET_COUNT_LIMIT_PER_MINER_INDEX = 3000000
 # How old a data entity bucket can be before the validators do not assign any value for them.
 DATA_ENTITY_BUCKET_AGE_LIMIT_DAYS = 30
-
-# The maximum length of a data label.
-MAX_DATA_LABEL_LENGTH = 32
-
-# The current protocol version.
-PROTOCOL_VERSION = 1
