+++ common/data.py
@@ -1,15 +1,10 @@
-from collections import defaultdict
 import dataclasses
-
 from common import constants
-from common.pydantic_dict_encoder import PydanticDictEncodersMixin
 from . import utils
 import datetime as dt
-from enum import Enum, IntEnum
-from typing import Any, Dict, List, Type, Optional
-from pydantic import (
-    BaseModel,
-)
+from enum import Enum, IntEnum, auto
+from typing import List, Type, Optional
+from pydantic import BaseModel, ConfigDict, Field, PositiveInt, validator
 
 
 class StrictBaseModel(BaseModel):
@@ -22,8 +17,6 @@ class StrictBaseModel(BaseModel):
         # enabling `use_enum_values`. It's possible this isn't an
         # issue with newer version of pydantic, which we can't use.
         use_enum_values = True
-
-        arbitrary_types_allowed = True
 
 
 @dataclasses.dataclass(frozen=True)
@@ -41,18 +34,15 @@ class DateRange:
         return self.start <= datetime < self.end
 
 
-@dataclasses.dataclass(frozen=True)
-class TimeBucket:
+class TimeBucket(StrictBaseModel):
     """Represents a specific time bucket in the linear flow of time."""
 
-    # Monotonically increasing value idenitifying the given time bucket
-    id: int
+    # Makes the object "Immutable" once created.
+    model_config = ConfigDict(frozen=True)
 
-    @classmethod
-    def todict(cls, v):
-        return {
-            "id": v.id,
-        }
+    id: PositiveInt = Field(
+        description="Monotonically increasing value idenitifying the given time bucket"
+    )
 
     @classmethod
     def from_datetime(cls, datetime: dt.datetime) -> Type["TimeBucket"]:
@@ -90,45 +80,32 @@ class DataSource(IntEnum):
     UNKNOWN_7 = 7
 
 
-@dataclasses.dataclass(frozen=True)
-class DataLabel:
+class DataLabel(StrictBaseModel):
     """An optional label to classify a data entity. Each data source will have its own definition and interpretation of labels.
 
     For example, in Reddit a label is the subreddit. For a stock price, it'll be the ticker symbol.
     """
 
-    value: str
+    class Config:
+        frozen = True
 
+    value: str = Field(
+        max_length=32,
+        description="The label. E.g. a subreddit for Reddit data.",
+    )
+
+    @validator("value")
     @classmethod
-    def todict(cls, v):
-        return {
-            "value": v.value,
-        }
-
-    def __getattribute__(self, prop):
-        if prop == "value":
-            v = super().__getattribute__(prop)
-            if v and len(v) > constants.MAX_DATA_LABEL_LENGTH:
-                raise ValueError(
-                    f"DataLabel cannot be longer than {constants.MAX_DATA_LABEL_LENGTH} characters."
-                )
-            return v.lower() if v else None
-        return super().__getattribute__(prop)
-
-    def __hash__(self):
-        return hash(self.value.lower() if self.value else None)
-
-    def __eq__(self, other):
-        if isinstance(other, DataLabel):
-            this_value = self.value.lower() if self.value else None
-            other_value = other.value.lower() if other.value else None
-            return this_value == other_value
-        return False
+    def lower_case_value(cls, value: str) -> str:
+        """Converts the value to lower case to consistent casing throughout the system."""
+        return value.lower()
 
 
-@dataclasses.dataclass(frozen=True)
-class DataEntity:
+class DataEntity(StrictBaseModel):
     """A logical unit of data that has been scraped. E.g. a Reddit post"""
+
+    # Makes the object "Immutable" once created.
+    model_config = ConfigDict(frozen=True)
 
     # Path from which the entity was generated.
     uri: str
@@ -136,9 +113,11 @@ class DataEntity:
     # Should be in UTC.
     datetime: dt.datetime
     source: DataSource
+    label: Optional[DataLabel] = Field(
+        default=None,
+    )
     content: bytes
-    content_size_bytes: int
-    label: Optional[DataLabel] = None
+    content_size_bytes: int = Field(ge=0)
 
     @classmethod
     def are_non_content_fields_equal(
@@ -152,37 +131,18 @@ class DataEntity:
             and this.label == other.label
         )
 
-    @classmethod
-    def todict(cls, v):
-        return {
-            "uri": v.uri,
-            "datetime": v.datetime.isoformat(),
-            "source": int(v.source),
-            "content": v.content,
-            "content_size_bytes": v.content_size_bytes,
-            "label": DataLabel.todict(v.label) if v.label else None,
-        }
 
-
-@dataclasses.dataclass(frozen=True)
-class DataEntityBucketId:
+class DataEntityBucketId(StrictBaseModel):
     """Uniquely identifies a bucket to group DataEntities by time bucket, source, and label."""
 
     time_bucket: TimeBucket
-    source: DataSource
-    label: Optional[DataLabel] = None
-
-    @classmethod
-    def todict(cls, v):
-        return {
-            "time_bucket": TimeBucket.todict(v.time_bucket),
-            "source": int(v.source),
-            "label": DataLabel.todict(v.label) if v.label else None,
-        }
+    source: DataSource = Field()
+    label: Optional[DataLabel] = Field(
+        default=None,
+    )
 
 
-@dataclasses.dataclass(frozen=True)
-class DataEntityBucket:
+class DataEntityBucket(StrictBaseModel):
     """Summarizes a group of data entities stored by a miner.
 
     Each bucket is uniquely identified by the time bucket, source, and label and it must be complete. i.e. a mine
@@ -191,133 +151,42 @@ class DataEntityBucket:
     A single bucket is limited to 128MBs to ensure requests sent over the network aren't too large.
     """
 
-    # Identifies the qualities by which this bucket is grouped.
-    id: DataEntityBucketId
-    size_bytes: int
-
-    @classmethod
-    def todict(cls, v):
-        return {
-            "id": DataEntityBucketId.todict(v.id),
-            "size_bytes": v.size_bytes,
-        }
+    id: DataEntityBucketId = Field(
+        description="Identifies the qualities by which this bucket is grouped."
+    )
+    size_bytes: int = Field(ge=0, le=constants.DATA_ENTITY_BUCKET_SIZE_LIMIT_BYTES)
 
 
-@dataclasses.dataclass(frozen=True)
-class ScorableDataEntityBucket:
+class ScorableDataEntityBucket(StrictBaseModel):
     """Composes both a DataEntityBucket and additional information required for scoring."""
 
-    # The DataEntityBucket that has additional information attached.
-    data_entity_bucket: DataEntityBucket
-
+    data_entity_bucket: DataEntityBucket = Field(
+        description="The DataEntityBucket that has additional information attached."
+    )
     # Scorable bytes are the bytes that can be credited to this miner for scoring.
     # This is always less than or equal to the total size of the chunk.
     # This scorable bytes are computed as:
     # 1 byte for every byte in size_bytes that no other miner has in their index.
     # 1 byte / # of miners that have this chunk in their index for every byte in size_bytes that at least one other miner has in their index.
-    scorable_bytes: int
+    scorable_bytes: int = Field(ge=0, le=constants.DATA_ENTITY_BUCKET_SIZE_LIMIT_BYTES)
 
 
-@dataclasses.dataclass(frozen=True)
-class MinerIndex:
+class MinerIndex(StrictBaseModel):
     """The Miner index."""
 
-    # ss58_address of the miner's hotkey.
-    hotkey: str
-
-    # Buckets the miner is serving.
-    data_entity_buckets: List[DataEntityBucket]
-
-    @classmethod
-    def compress(cls, index: "MinerIndex") -> "CompressedMinerIndex":
-        """Compresses a MinerIndex to reduce bytes sent on the wire."""
-        sources_to_buckets = defaultdict(dict)
-
-        # Iterate through the buckets and aggregate them by source and label.
-        for bucket in index.data_entity_buckets:
-            labels_to_buckets = sources_to_buckets[bucket.id.source]
-
-            # The label may be None, but this is a valid key in a dict.
-            if bucket.id.label in labels_to_buckets:
-                labels_to_buckets[bucket.id.label].time_bucket_ids.append(
-                    bucket.id.time_bucket.id
-                )
-                labels_to_buckets[bucket.id.label].sizes_bytes.append(bucket.size_bytes)
-            else:
-                labels_to_buckets[bucket.id.label] = CompressedEntityBucket(
-                    label=bucket.id.label,
-                    time_bucket_ids=[bucket.id.time_bucket.id],
-                    sizes_bytes=[bucket.size_bytes],
-                )
-
-        return CompressedMinerIndex(
-            sources={
-                source: list(labels_to_buckets.values())
-                for source, labels_to_buckets in sources_to_buckets.items()
-            },
-        )
+    hotkey: str = Field(min_length=1, description="ss58_address of the miner's hotkey.")
+    data_entity_buckets: List[DataEntityBucket] = Field(
+        description="Buckets the miner is serving.",
+        max_items=constants.DATA_ENTITY_BUCKET_COUNT_LIMIT_PER_MINER_INDEX,
+    )
 
 
-@dataclasses.dataclass(frozen=True)
-class ScorableMinerIndex:
+class ScorableMinerIndex(StrictBaseModel):
     """The Miner index, with additional information required for scoring."""
 
-    # ss58_address of the miner's hotkey.
-    hotkey: str
-
-    # DataEntityBuckets the miner is serving, scored on uniqueness.
-    scorable_data_entity_buckets: List[ScorableDataEntityBucket]
-
-    # Time last updated in UTC.
-    last_updated: dt.datetime
-
-
-@dataclasses.dataclass()
-class CompressedEntityBucket:
-    """A compressed version of the DataEntityBucket to reduce bytes sent on the wire."""
-
-    # The label of the bucket.
-    label: Optional[DataLabel] = None
-
-    # A list of time bucket ids for this label and source. Must match the length of sizes_bytes.
-    time_bucket_ids: List[int] = dataclasses.field(
-        default_factory=list,
+    hotkey: str = Field(min_length=1, description="ss58_address of the miner's hotkey.")
+    scorable_data_entity_buckets: List[ScorableDataEntityBucket] = Field(
+        description="DataEntityBuckets the miner is serving, scored on uniqueness.",
+        max_items=constants.DATA_ENTITY_BUCKET_COUNT_LIMIT_PER_MINER_INDEX,
     )
-
-    # A list of sizes in bytes for each time bucket where sizes_bytes[i] is the size_bytes for time_bucket_ids[i].",
-    sizes_bytes: List[int] = dataclasses.field(
-        default_factory=list,
-    )
-
-
-@dataclasses.dataclass(frozen=True)
-class CompressedMinerIndex:
-    """A compressed version of the MinerIndex to reduce bytes sent on the wire."""
-
-    # A map from source to a list of compressed buckets.
-    sources: Dict[DataSource, List[CompressedEntityBucket]]
-
-    @classmethod
-    def decompress(
-        cls, compressed_index: "CompressedMinerIndex", hotkey: str
-    ) -> MinerIndex:
-        """Decompresses a compressed index."""
-        data_entity_buckets = []
-
-        for source, compressed_buckets in compressed_index.sources.items():
-            for compressed_bucket in compressed_buckets:
-                for time_bucket_id, size_bytes in zip(
-                    compressed_bucket.time_bucket_ids, compressed_bucket.sizes_bytes
-                ):
-                    data_entity_buckets.append(
-                        DataEntityBucket(
-                            id=DataEntityBucketId(
-                                time_bucket=TimeBucket(id=time_bucket_id),
-                                source=source,
-                                label=compressed_bucket.label,
-                            ),
-                            size_bytes=size_bytes,
-                        )
-                    )
-
-        return MinerIndex(hotkey=hotkey, data_entity_buckets=data_entity_buckets)
+    last_updated: dt.datetime = Field(description="Time last updated in UTC.")
