+++ tests/vali_utils/test_validator_utils.py
@@ -1,16 +1,10 @@
-from calendar import c
-import bittensor as bt
 import datetime as dt
 import unittest
-from common import constants
 from common.data import (
-    CompressedEntityBucket,
-    CompressedMinerIndex,
     DataEntityBucket,
     DataEntity,
     DataEntityBucketId,
     DataLabel,
-    MinerIndex,
     TimeBucket,
     DataSource,
     ScorableDataEntityBucket,
@@ -21,7 +15,7 @@ import vali_utils.utils as vali_utils
 import pytz
 
 
-class TestValiUtils(unittest.TestCase):
+class TestValidatorUtils(unittest.TestCase):
     def test_choose_data_entity_bucket_to_query(self):
         """Calls choose_data_entity_bucket_to_query 10000 times and ensures the distribution of bucketss chosen is as expected."""
         index = ScorableMinerIndex(
@@ -464,184 +458,6 @@ class TestValiUtils(unittest.TestCase):
         ]
         unique = vali_utils.are_entities_unique(entities)
         self.assertFalse(unique)
-
-    def test_get_miner_index_from_response_old_index(self):
-        """Tests get_miner_index_from_response with an old index."""
-
-        buckets = [
-            DataEntityBucket(
-                id=DataEntityBucketId(
-                    time_bucket=TimeBucket(id=5),
-                    source=DataSource.REDDIT,
-                    label=DataLabel(value="r/bittensor_"),
-                ),
-                size_bytes=100,
-            ),
-            DataEntityBucket(
-                id=DataEntityBucketId(
-                    time_bucket=TimeBucket(id=6),
-                    source=DataSource.X,
-                    label=DataLabel(value="#bittensor"),
-                ),
-                size_bytes=200,
-            ),
-        ]
-        response = GetMinerIndex(
-            data_entity_buckets=buckets, dendrite=bt.TerminalInfo(status_code=200)
-        )
-
-        index = vali_utils.get_miner_index_from_response(response, "hk")
-        expected_index = MinerIndex(hotkey="hk", data_entity_buckets=buckets)
-        self.assertEqual(index, expected_index)
-
-    def test_get_miner_index_from_response_old_index_bucket_size_too_large(self):
-        """Tests get_miner_index_from_response with an old index."""
-
-        buckets = [
-            DataEntityBucket(
-                id=DataEntityBucketId(
-                    time_bucket=TimeBucket(id=5),
-                    source=DataSource.REDDIT,
-                    label=DataLabel(value="r/bittensor_"),
-                ),
-                size_bytes=100,
-            ),
-            DataEntityBucket(
-                id=DataEntityBucketId(
-                    time_bucket=TimeBucket(id=6),
-                    source=DataSource.X,
-                    label=DataLabel(value="#bittensor"),
-                ),
-                size_bytes=constants.DATA_ENTITY_BUCKET_SIZE_LIMIT_BYTES + 1,
-            ),
-        ]
-        response = GetMinerIndex(
-            data_entity_buckets=buckets, dendrite=bt.TerminalInfo(status_code=200)
-        )
-
-        with self.assertRaises(ValueError):
-            vali_utils.get_miner_index_from_response(response, "hk")
-
-    def test_get_miner_index_from_response_compressed_index(self):
-        """Tests get_miner_index_from_response with a compressed index."""
-
-        compressed_index = CompressedMinerIndex(
-            sources={
-                DataSource.REDDIT: [
-                    CompressedEntityBucket(
-                        label=DataLabel(value="r/bittensor_"),
-                        time_bucket_ids=[5, 6],
-                        sizes_bytes=[100, 200],
-                    )
-                ],
-                DataSource.X: [
-                    CompressedEntityBucket(
-                        label=DataLabel(value="#bittensor"),
-                        time_bucket_ids=[6],
-                        sizes_bytes=[300],
-                    )
-                ],
-            }
-        )
-
-        response = GetMinerIndex(
-            compressed_index=compressed_index, dendrite=bt.TerminalInfo(status_code=200)
-        )
-
-        index = vali_utils.get_miner_index_from_response(response, "hk")
-        expected_index = MinerIndex(
-            hotkey="hk",
-            data_entity_buckets=[
-                DataEntityBucket(
-                    id=DataEntityBucketId(
-                        time_bucket=TimeBucket(id=5),
-                        source=DataSource.REDDIT,
-                        label=DataLabel(value="r/bittensor_"),
-                    ),
-                    size_bytes=100,
-                ),
-                DataEntityBucket(
-                    id=DataEntityBucketId(
-                        time_bucket=TimeBucket(id=6),
-                        source=DataSource.REDDIT,
-                        label=DataLabel(value="r/bittensor_"),
-                    ),
-                    size_bytes=200,
-                ),
-                DataEntityBucket(
-                    id=DataEntityBucketId(
-                        time_bucket=TimeBucket(id=6),
-                        source=DataSource.X,
-                        label=DataLabel(value="#bittensor"),
-                    ),
-                    size_bytes=300,
-                ),
-            ],
-        )
-        self.assertEqual(index, expected_index)
-
-    def test_get_miner_index_from_response_new_index_bucket_size_too_large(self):
-        """Tests get_miner_index_from_response with a compressed index that has a bucket that is too large."""
-
-        compressed_index = CompressedMinerIndex(
-            sources={
-                DataSource.REDDIT: [
-                    CompressedEntityBucket(
-                        label=DataLabel(value="r/bittensor_"),
-                        time_bucket_ids=[5, 6],
-                        sizes_bytes=[100, 200],
-                    )
-                ],
-                DataSource.X: [
-                    CompressedEntityBucket(
-                        label=DataLabel(value="#bittensor"),
-                        time_bucket_ids=[6],
-                        sizes_bytes=[constants.DATA_ENTITY_BUCKET_SIZE_LIMIT_BYTES + 1],
-                    )
-                ],
-            }
-        )
-
-        response = GetMinerIndex(
-            compressed_index=compressed_index, dendrite=bt.TerminalInfo(status_code=200)
-        )
-
-        with self.assertRaises(ValueError):
-            vali_utils.get_miner_index_from_response(response, "hk")
-
-    def test_get_miner_index_from_response_new_index_too_many_buckets(self):
-        """Tests get_miner_index_from_response with an old index."""
-
-        compressed_index = CompressedMinerIndex(
-            sources={
-                DataSource.REDDIT: [
-                    CompressedEntityBucket(
-                        label=DataLabel(value="r/bittensor_"),
-                        time_bucket_ids=[
-                            i
-                            for i in range(
-                                constants.DATA_ENTITY_BUCKET_COUNT_LIMIT_PER_MINER_INDEX
-                                + 1
-                            )
-                        ],
-                        sizes_bytes=[
-                            i
-                            for i in range(
-                                constants.DATA_ENTITY_BUCKET_COUNT_LIMIT_PER_MINER_INDEX
-                                + 1
-                            )
-                        ],
-                    )
-                ],
-            }
-        )
-
-        response = GetMinerIndex(
-            compressed_index=compressed_index, dendrite=bt.TerminalInfo(status_code=200)
-        )
-
-        with self.assertRaises(ValueError):
-            vali_utils.get_miner_index_from_response(response, "hk")
 
 
 if __name__ == "__main__":
